#EA-2340 added pagination for closed sets
