+++ set-web/src/main/java/eu/europeana/set/web/service/UserSetService.java
@@ -1,5 +1,13 @@
 package eu.europeana.set.web.service;
 
+import java.io.IOException;
+import java.util.List;
+
+import javax.servlet.http.HttpServletRequest;
+
+import org.codehaus.jettison.json.JSONException;
+import org.springframework.security.core.Authentication;
+
 import eu.europeana.api.commons.definitions.search.ResultSet;
 import eu.europeana.api.commons.web.exception.ApplicationAuthenticationException;
 import eu.europeana.api.commons.web.exception.HttpException;
@@ -11,15 +19,8 @@ import eu.europeana.set.mongo.model.internal.PersistentUserSet;
 import eu.europeana.set.web.exception.request.RequestBodyValidationException;
 import eu.europeana.set.web.exception.response.UserSetNotFoundException;
 import eu.europeana.set.web.model.search.BaseUserSetResultPage;
+import eu.europeana.set.web.model.search.CollectionPage;
 import eu.europeana.set.web.model.search.ItemIdsResultPage;
-
-import org.codehaus.jettison.json.JSONException;
-import org.springframework.security.core.Authentication;
-
-import java.io.IOException;
-import java.util.List;
-
-import javax.servlet.http.HttpServletRequest;
 
 public interface UserSetService {
 
@@ -179,11 +180,13 @@ public interface UserSetService {
 	    Authentication authentication);
 
     public BaseUserSetResultPage<?> buildResultsPage(UserSetQuery searchQuery, ResultSet<? extends UserSet> results,
-	    StringBuilder requestUrl, String reqParams, LdProfiles profile, Authentication authentication)
+	    String requestUrl, String reqParams, LdProfiles profile, Authentication authentication)
 	    throws HttpException;
     
     public ItemIdsResultPage buildItemIdsResultsPage(List<String> itemIds, int page, int pageSize,
 	    HttpServletRequest request);
+    
+    public CollectionPage buildCollectionPage(UserSet userSet, LdProfiles profile, int pageNr, int pageSize, HttpServletRequest request) throws HttpException;
 
     /**
      * This method validates input values wsKey, identifier and userToken.
