+++ set-web/src/main/java/eu/europeana/set/web/service/controller/BaseRest.java
@@ -28,6 +28,7 @@ import eu.europeana.set.definitions.exception.UserSetProfileValidationException;
 import eu.europeana.set.definitions.model.UserSet;
 import eu.europeana.set.definitions.model.vocabulary.LdProfiles;
 import eu.europeana.set.web.http.UserSetHttpHeaders;
+import eu.europeana.set.web.model.search.CollectionPage;
 import eu.europeana.set.web.search.UserSetLdSerializer;
 import eu.europeana.set.web.service.UserSetService;
 import eu.europeana.set.web.service.authorization.UserSetAuthorizationService;
@@ -160,6 +161,14 @@ public class BaseRest extends BaseRestController {
 	return serializer.serialize(set);
     }
     
+    protected String serializeCollectionPage(CollectionPage itemPage) throws IOException {
+	//prepare data for serialization according to the profile
+	
+	UserSetLdSerializer serializer = new UserSetLdSerializer();
+	return serializer.serialize(itemPage);
+    }
+    
+    
     protected String serializeResultPage(LdProfiles profile, UserSet storedUserSet) throws IOException {
 	//prepare data for serialization according to the profile
 	UserSet set = getUserSetService().applyProfile(storedUserSet, profile);
@@ -197,8 +206,16 @@ public class BaseRest extends BaseRestController {
         return getConfiguration().getApiVersion();
     }
     
-    protected ResponseEntity<String> buildGetResponse(UserSet userSet, LdProfiles profile) throws IOException {
-	String userSetJsonLdStr = serializeUserSet(profile, userSet);
+    protected ResponseEntity<String> buildGetResponse(UserSet userSet, LdProfiles profile, int pageNr, int pageSize, HttpServletRequest request) throws IOException, HttpException {
+	String jsonBody = "";
+	if(pageNr < 0) {
+	    jsonBody = serializeUserSet(profile, userSet);    
+	}else {
+	    CollectionPage itemPage = getUserSetService().buildCollectionPage(userSet, profile, pageNr, pageSize, request);
+	    jsonBody = serializeCollectionPage(itemPage);
+	}
+	
+	
 	String etag = generateETag(userSet.getModified(), WebFields.FORMAT_JSONLD, getApiVersion());
 
 	// build response
@@ -211,7 +228,7 @@ public class BaseRest extends BaseRestController {
 	// generate “ETag”;
 	headers.add(UserSetHttpHeaders.ETAG, etag);
 
-	return new ResponseEntity<>(userSetJsonLdStr, headers, HttpStatus.OK);
+	return new ResponseEntity<>(jsonBody, headers, HttpStatus.OK);
     }
 
 }
